+++ basics/processing-instructions/steel/api/src/lib.rs
@@ -1,6 +1,8 @@
+pub mod instruction;
 pub mod sdk;
 
 pub mod prelude {
+    pub use crate::instruction::*;
     pub use crate::sdk::*;
 }
