+++ basics/processing-instructions/steel/program/src/lib.rs
@@ -1,23 +1,27 @@
+mod go_to_the_park;
+
+use go_to_the_park::*;
 use processing_instructions_api::prelude::*;
 use solana_program::msg;
 use steel::*;
 
 pub fn process_instruction(
-    _program_id: &Pubkey,
-    _accounts: &[AccountInfo],
+    program_id: &Pubkey,
+    accounts: &[AccountInfo],
     instruction_data: &[u8],
 ) -> ProgramResult {
     msg!("Processing instruction");
 
-    let instruction = GoToTheParkData::try_from_bytes(instruction_data)?;
+    let (ix, data) = parse_instruction(
+        &processing_instructions_api::ID,
+        program_id,
+        instruction_data,
+    )?;
 
-    msg!("Welcome to the park, {}!", instruction.name());
+    match ix {
+        ProcessingInstructionsInstruction::GoToThePark => process_go_to_the_park(accounts, data)?,
+    }
 
-    if instruction.height() > 5 {
-        msg!("You are tall enough to ride this ride. Congratulations.");
-    } else {
-        msg!("You are NOT tall enough to ride this ride. Sorry mate.");
-    };
     Ok(())
 }
